Deletion: BlockExchangePacket
